+++ exercises/prime-factors/PrimeFactors.cs
@@ -2,7 +2,7 @@ using System;
 
 public static class PrimeFactors
 {
-    public static int[] Factors(long number)
+    public static long[] Factors(long number)
     {
         throw new NotImplementedException();
     }
